converted float to int (cents) and wrote a function to convert cents to dollars with optional currency sign second parameter

+++ server/db/models/product.js
@@ -11,8 +11,9 @@ const Product = db.define('product', {
     type: Sequelize.TEXT,
     allowNull: false
   },
+  //  stored in cents
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false
   },
   image: {
